Implement Stable Kaldi IOS project.
	modified:   src/base/kaldi-math.h
	modified:   src/feat/feature-mfcc.cc
	modified:   src/feat/mel-computations.cc
	modified:   src/feat/wave-reader.cc
	modified:   src/featbin/Makefile
	modified:   src/featbin/compute-mfcc-feats.cc
	modified:   src/fstext/context-fst.h
	modified:   src/fstext/determinize-star-inl.h
	modified:   src/fstext/trivial-factor-weight.h
	modified:   src/lat/determinize-lattice-pruned.cc
	modified:   src/lm/kaldi-lmtable.h
	modified:   src/nnet2/combine-nnet-fast.cc
	modified:   src/nnet2/combine-nnet.cc
	modified:   src/online2bin/Makefile
	modified:   src/util/stl-utils.h
	new file:   stable-kaldi/stable-kaldi.xcodeproj/project.pbxproj
	new file:   stable-kaldi/stable-kaldi.xcodeproj/project.xcworkspace/contents.xcworkspacedata
	new file:   stable-kaldi/stable-kaldi.xcodeproj/project.xcworkspace/xcuserdata/kevinyang.xcuserdatad/UserInterfaceState.xcuserstate
	new file:   stable-kaldi/stable-kaldi.xcodeproj/xcuserdata/kevinyang.xcuserdatad/xcschemes/stable-kaldi.xcscheme
	new file:   stable-kaldi/stable-kaldi.xcodeproj/xcuserdata/kevinyang.xcuserdatad/xcschemes/xcschememanagement.plist

+++ src/feat/wave-reader.cc
@@ -220,7 +220,7 @@ void WaveData::Read(std::istream &is) {
   uint32 data_chunk_size = ReadUint32(is, swap);
   riff_chunk_read += 4;
 
-  if (std::abs((riff_chunk_read + data_chunk_size) - riff_chunk_size) > 1) {
+  if (std::abs(static_cast<int>((riff_chunk_read + data_chunk_size) - riff_chunk_size)) > 1) {
     // we allow the size to be off by one, because there is a weirdness in the
     // format of RIFF files that means that the input may sometimes be padded
     // with 1 unused byte to make the total size even.
